Made substance Id long

+++ src/main/java/com/dpw/runner/shipment/services/service/impl/PackingService.java
@@ -149,7 +149,7 @@ public class PackingService implements IPackingService {
     }
 
     private static void valiadteDGSubstanceId(Map<Long, Long> dicDGSubstanceUNDGContact, int row, Packing packingRow) {
-        if (packingRow.getDGSubstanceId() != null && !dicDGSubstanceUNDGContact.containsKey(packingRow.getDGSubstanceId())) {
+        if (packingRow.getDGSubstanceId() != null && !dicDGSubstanceUNDGContact.containsKey(packingRow.getDGSubstanceId().longValue())) {
             throw new ValidationException("DG Substance Id is invalid at row: " + row);
         }
     }
